% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_function.R
\name{.pt_agrep_whole}
\alias{.pt_agrep_whole}
\title{Function to match the closest fuzzy name}
\usage{
.pt_agrep_whole(x, y, max_distance)
}
\description{
Function to match the closest fuzzy name
}
\keyword{internal}
