% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_algorithm.R
\name{.pt_exact_match}
\alias{.pt_exact_match}
\title{Exact match function}
\usage{
.pt_exact_match(splist_class_i, pos_genus, n_class, fuzzy = FALSE)
}
\description{
Exact match function
}
\keyword{internal}
