% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_algorithm.R
\name{.pt_fuzzy_match}
\alias{.pt_fuzzy_match}
\title{Fuzzy matching function}
\usage{
.pt_fuzzy_match(splist_class_i, pos_genus = NULL, max.distance, n_class)
}
\description{
Fuzzy matching function
}
\keyword{internal}
