% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searh_group.R
\name{.pt_group}
\alias{.pt_group}
\title{Function to wrap .pt_group_ind for multiple names}
\usage{
.pt_group(group_names, group_ref, max_distance)
}
\description{
Function to wrap .pt_group_ind for multiple names
}
\keyword{internal}
