% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searh_group.R
\name{.pt_group_ind}
\alias{.pt_group_ind}
\title{Function to search species names,
based on group (genus, family, order)}
\usage{
.pt_group_ind(
  group_name,
  group_ref,
  max_distance,
  only_one = TRUE,
  closest = FALSE
)
}
\description{
Function to search species names,
based on group (genus, family, order)
}
\keyword{internal}
