% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_algorithm.R
\name{.pt_match_algorithm}
\alias{.pt_match_algorithm}
\title{Code for the algorithms to exactly and fuzzy match names
Author: Bruno Vilela
The matching algorithm}
\usage{
.pt_match_algorithm(splist_class, max.distance)
}
\description{
Code for the algorithms to exactly and fuzzy match names
Author: Bruno Vilela
The matching algorithm
}
\keyword{internal}
