% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_function.R
\name{.pt_names_check}
\alias{.pt_names_check}
\title{Function to check list of names input}
\usage{
.pt_names_check(splist, argument_name)
}
\description{
Function to check list of names input
}
\keyword{internal}
