% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_function.R
\name{.pt_splist_classify}
\alias{.pt_splist_classify}
\title{This code has two functions to classify species names according to
Genus, epithet, authors and infracategories
Function wrap of .classify_algo for multiple species}
\usage{
.pt_splist_classify(x)
}
\description{
This code has two functions to classify species names according to
Genus, epithet, authors and infracategories
Function wrap of .classify_algo for multiple species
}
\keyword{internal}
