% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perutimber-package.R
\docType{package}
\name{perutimber-package}
\alias{perutimber}
\alias{perutimber-package}
\title{perutimber: Catalogue Of The Timber Forest Species Of The Amazon And The Peruvian Yunga}
\description{
The data of the Catalogue Of The Timber Forest Species Of The Amazon And The Peruvian Yunga.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/perutimber}
  \item \url{https://paulesantos.github.io/perutimber/}
  \item Report bugs at \url{https://github.com/PaulESantos/perutimber/issues}
}

}
\author{
\strong{Maintainer}: Paul Efren Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID})

}
\keyword{internal}
