% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pt.R
\docType{data}
\name{perutimber_data}
\alias{perutimber_data}
\title{List of Plant Species Name and information from the Catalogue of the timber forest
species of the Amazon and the Peruvian Yunga.}
\format{
A tibble with the following columns:
\describe{\item{accepted_name}{A character vector. The list of the accepted plant taxa names according to the Taxonomic Name Resolution Service - TNRS.}
\item{accepted_name_author}{A character vector. The corresponding author name of the accepted_name, staying empty if the taxonomic_status is "Synonym" or "No opinion".}
\item{accepted_family}{A character vector. The corresponding family name of the accepted_name.}
\item{common_name}{A character vector. Listed all common names for each species.}
\item{accepted_name_rank}{A character vector. Information about infraspecific status.}
\item{habit}{A character vector.}
\item{plant_height_m}{A character vector. The plant height range in meters.}
\item{regions}{A character vector.}
\item{elevation}{A character vector.}
\item{uses}{A character vector.}}
}
\usage{
perutimber_data
}
\description{
The 'perutimber_dat' contains information from the Catalogue of the timber forest
species of the Amazon and the Peruvian Yunga.
}
\examples{

data(perutimber_data)
str(perutimber_data)
}
\keyword{datasets}
