% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pt.R
\docType{data}
\name{perutimber_sps_class}
\alias{perutimber_sps_class}
\title{List of species name in tab_perutimber separeted by category}
\format{
A data.frame.
}
\usage{
perutimber_sps_class
}
\description{
The 'perutimber_sps_class' includes all species separeted by genus, epithet, author,
subspecies, variety, and id (position in the
\code{tab_perutimber}).
}
\examples{

data(perutimber_sps_class)

}
\keyword{datasets}
