% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pt.R
\docType{data}
\name{tab_perutimber_position}
\alias{tab_perutimber_position}
\title{List of the number positions of the first 3 letters of the species name in
the tab_perutimber}
\format{
A data frame with 399 observations on the following 3 variables.
\describe{
\item{position}{A character vector. It is the position of
the first 3 letters of the species name in the tab_perutimber.}
\item{triphthong}{A character vector. First 3 letters of the species
name in the tab_perutimber.}
\item{genus}{A character vector.
Corresponding Genus name.} }
}
\usage{
tab_perutimber_position
}
\description{
The 'tab_perutimber_position' reports the
position (in term of number of rows) of the first three letters (triphthong)
for the plant names stored in the variable 'accepted_name' of the table
'tab_perutimber'. This indexing system speeds up of the search on the
largest list using the package.
}
\examples{

data(tab_perutimber_position)
str(tab_perutimber_position)
}
\keyword{datasets}
