% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perutimber-package.R
\docType{package}
\name{perutimber-package}
\alias{perutimber}
\alias{perutimber-package}
\title{perutimber: Catalogue of the Timber Forest Species of the Peruvian Amazon}
\description{
Access the data of the 'Catalogue of the Timber Forest Species of the Peruvian Amazon' Vásquez Martínez, R., & Rojas Gonzáles, R.D.P.(2022)\doi{10.21704/rfp.v37i3.1956}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/perutimber}
  \item \url{https://paulesantos.github.io/perutimber/}
  \item Report bugs at \url{https://github.com/PaulESantos/perutimber/issues}
}

}
\author{
\strong{Maintainer}: Paul E. Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID})

}
\keyword{internal}
